Add required HTML5 attribute to some input fields
The `required` attribute is a less destructive way of telling a
user that a form field cannot be left blank than redirecting to an
error page (which usually results in losing any form fields already
filled in and having to navigate back to the page).

It can easily be circumvented (by falsifying form data or by not
having an HTML5-compatible browser), so the checks on these fields
in the PHP need to remain.

However, it is still very useful, so I suspect I'll be adding it
more systematically throughout the code later.