Add caching directive for static content
Images and other static assets that are currently not cache-busted
will be cached for a week.

CSS and JS files (which we currently cache-bust) can be cached for
an arbitrarily long time, so we set it to a year.

This change fixes F grades on performance test sites in the
"Cache static content" category. Now we get an A!

NOTE: No particularly strong reason why `mod_headers` was used to set
the "max-age" header instead of `mod_expires`.