SmrPlayer: remove `get` method
This method duplicated all our getter methods, but using array keys
to decide which property to access. This was briefly used in 2014
(8756979d839cf), but was quickly changed.

Since this is not the way we want class access to work, we simply
remove this unused method to avoid it from being used in the future.