{ship,weapon}_list.php: convert to list.js
...and other simplifications.

Use list.js to sort these tables instead of doing the processing
on the server. This is much faster for the user. We also include
a seach box.

We also massively simplify the javascript for filtering by passing
the element we're filtering. This allows us to remove all of the
hard-coded "picker" wrappers (e.g. `speedPickf`, 'racePickf`, etc).
As a result, we can reuse the same javascript file for both pages,
which we rename to `filter_list.js`.

Move most of the code that builds the `<select>` choices from the
engine file to the templates. Note also that `<option>` does not
need a `value` if it is the same as the text of the option.

Use `SmrShip::getAllBaseShips` instead of manually constructing
the query. This is presently a bit inefficient because this method
is not fully optimized (we get an extra query for each ship).
This *can* be fixed, but for now the performance impact is not
worth the extra complexity.