Fix script tags breaking tables in chrome
This fix prevents jitter in pages that included list.js in the middle
of the HTML body.

While it is preferable to include the list.js code in the same file
as the table we are sorting (since there are table classes that must
exactly match JS variables), including `<script>` elements in the
middle of the HTML body seems to break the rendering in Chrome. This
issue manifests as a quick flash when the page first loads where the
content after the `<script>` is missing, creating the appearance of
jitter in the page elements, which can be quite distracting.

It is a best practice to include `<script>` tags at the end of the
body (or in `<head>`), but it shouldn't be required. In fact, Firefox
has no problem with the existing code. However, since Chrome is so
ubiquitous, we will work around this apparent bug by adding a hook to
load the list.js content after the body content.

The list.js code has been moved to a separate JS file (and refactored
to avoid code duplication) in a `listjs` namespace. Then in the body
of the template, we call the new function `Template->setListjsInclude`
which then directs `EndingJavascript.inc` to call this function.

A few `data-X` attributes were added to fix the sorting of player
names when their rank was prefixed to the name we actually want to
sort on.