Add a hook in Template to load a javascript file
The new method `Template::addJavascriptSource` is the hook between
the template body and the EndingJavascript.inc (JS injected into the
bottom of the body).

This method populates a `$jsSources` array, which should give the
user-facing path to the JS file to source.

NOTE: `addJavascriptSource` is private so that it is only called
from within the template body. If it were called in the engine files,
then those paths would not be cache-busted (since we currently only
cache-bust in `templates` and `htdocs`).