Move JS functions from body script to smr15.js
In shop_hardware.php, this allows us to simply remove the `<script>`
from the body of the template.

For planet_defense.php, we need to tweak how we trigger the JS
function. We use the `onchange` attribute of the `<select>` element,
which replaces the need to install the 'change' handler for the
'div-toggle' class. The function body itself can be reused verbatim.