shop_goods_trade.php: use autofocus instead of JS
The HTML5 attribute `autofocus` does the same thing as the `focus()`
method in Javascript. We add this to the "bargain price" input, which
allows us to remove the `<script>` in the template body.

NOTE: We are no longer highlighting the bargain price (`select()`),
but focusing is sufficient to allow hitting "Enter" when the page
loads, which is all we need here.