SmrPlayer: remove NPC label from getPlayerName
This function is sometimes used in data contexts, rather than in
display contexts, so it should not contain any HTML elements.

The specific problem that prompted this change was the `data-name`
attribute in the Current Players list.js attribute.

NOTE: We could have instead escaped or stripped the HTML elements
in the places where it is expected to be used as data, but I think
it's just cleaner to remove the tag altogether. If it's confusing
not to have the `[NPC]` tag somewhere, we can either re-add it or
replace that usage with `getDisplayName`.