Modify how alliance roles are set for players
Use the `joinAlliance` method for all alliance joining (both for the
Leader who creates the alliance and any New Members). Then we make
the `setAllianceID` method protected.

Inserting the `player_has_alliance_role` for the Leader will now be
done in `joinAlliance` instead of `SmrAlliance::createDefaultRoles`,
for consistency with how it is done for New Members.

Also add global constants for the numeric values of the "Leader" and
"New Member" role ID's.