Refactor course plot processing
Since NPC's call the conventional and "plot to nearest" processing
pages, those pages need to set the plotted course before forwarding,
otherwise those containers will do nothing and the NPC will break.

Setting the plotted course was moved out of the processing page and
into the "results" page in c604fba873e because it was a routine that
needed to be run by both processing page, and the "results" page was
a common place to do this.

Now, we instead move this common processing into a separate file to
be called by both processing pages: `course_plot_processing.inc`.
This leaves the "results" page as a purely display page.

Now that we are no longer modifying the `$path` in the "results" page,
we also no longer need to (de)serialize it.