ajax.js: fix multiple actions errors
The JS code to prevent multiple actions omitted the case of form
submission, which meant that multiple form submissions or form
submissions before or after a link click / hotkey press would
result in the "Multiple actions..." error.

The most commonly reported scenario where this occurred was when
players clicked on Buy / Sell while trading, and then hit enter to
accept the bargain before the next page had actually loaded (which
occurs with some regularity due to the repetitive nature of trading).

To fix this, we add a submit hook on all form elements to check if
any other action has already been taken, and if not, to broadcast
that an action was taken. We use the shared `linkFollowed` variable
to achieve this.