Raise exceptions on PHP warnings in beta mode
Since people are unlikely to report issues they're seeing in beta,
and since standard warnings in the logs do not provide enough info
to debug where they are coming from, convert them into errors.

This will achieve two things:

1. It will make beta testers more aware of the issue.
2. It will send an e-mail with the full diagnostic details that
   we get when an exception in uncaught on the live server.

Note that we don't install this warnings-as-errors handler in
debug mode, since we generally want to display issues in the context
of the page using xdebug's handler.