npc.php: avoid a warning about headers after output
If we output before the SmrSession is initialized, then when
`SmrSession::init` calls `setcookie`, it will warn that we can't
send headers after output has already begun.

Since npc.php is a CLI script, headers don't matter anyway, but
we want to avoid the warning (because it is set to raise on beta).

Another way to avoid this warning would be to skip `setcookie` if
`NPC_SCRIPT` is defined, but it seemed more hacky to complicate
the logic in SmrSession.