npc.php: choose npc's by last_turn_update
Selecting an NPC by highest number of turns will not work correctly
because that does not factor in the turns gained since the last turn
update.

We could in theory create every NPC SmrPlayer, calculate their up-
to-date turn count, and then select from there, but that is a lot
of work for something that doesn't really matter anyway.

Picking the NPC that has acted least recently seems reasonable enough.