npc.php: remove NMV from ship upgrade logic
We have designed NPC's to be veterans when they join the game, so
they will no longer ever have a NMV (in fact, since we don't specify
a ship to give them, they should start with a Galactic Semi).
Therefore we don't need any NMV casing in the ship upgrade logic.