Optimize galaxy-wide database queries
We have a lot of methods that get an entity for a given sector
by querying the database if it is the first time the entity has
been requested. If we're performing a separate query for each
sector in a galaxy, then the database communication overhead
becomes the dominant cost.

On these pages, we can mitigate this overhead cost by performing
a single query for the entire galaxy to cache all the entities
that will be needed later.

The caching infrastructure was already set up for the most part,
we just needed methods to operate on a galaxy.

Adds galaxy-wide static caching methods to the following classes:

* SmrForce
* SmrPlanet
* SmrPlayer
* SmrPort
* SmrLocation

These are a bit more complicated than `SmrSector::getGalaxySectors`
because they only conditionally exist, but the principle is the
same.

These new methods are then used to pre-cache for the following pages:

* Galaxy Map
* Universe Generator
* DL Sectors File

NOTE: We remove the cache clearing in SectorMap.inc, which will
significantly increase concurrent memory usage, but for now we'll
trade off memory for speed.