universe_create_sectors.php: sanitize containers
By using `$container = $var`, we were accidentally populating the
links with unnecessary container variables. Instead, we now construct
the containers from scratch and only populate them with the variables
we need.

The effect of this was very subtle. Because the containers were given
additional variables (in this case, a `sector_id`, which was varying),
it was causing the CommonID -- and thus also the SN -- of the
containers to change.

This had two consequences:

1. There was a lot of bloat in the number of CommonIDs, and caused
   containers from being reused in certain cases.

2. In the following patch that adds ajax for modifying connections,
   the SN's changing for the `move_hack` link meant that every sector
   div was being rewritten by javascript instead of just the two
   sectors whose connections were changing.

As far as I can tell, these were merely inefficiencies, but it was
helpful to work out (and fix) what was causing them.