universe_create_sectors.php: modify walls with ajax
Galaxy maps are very large, and the process of manually adjusting
sector connections was slowed down by the fact that adding/removing
a wall reset the page display to the top left.

By using ajax, the process is much faster because the position on the
page remains unchanged. It also reduces bandwidth because only a small
amount of data on the page is modified, and so we don't need to serve
the entire page each time.