Uni Gen: add galaxy details editing
This new option allows for modifying galaxy metadata (name, type,
max force time) as well as galaxy dimensions.

Modifying dimensions is messy because it needs to change the primary
keys for a number of tables, and primary keys are not intended to be
changed.

When modifying dimensions, the algorithm will do the best it can
to retain the existing warps, locations, connections, etc. in the
grid-location they were in before (but obviously the sector IDs
will have changed).

Modifying dimensions cannot be used on enabled games.