Fix voting links in new web browsers
Due to a security fix in Chrome 76 and Firefox 70, it is no longer
possible to execute javascript in a new tab. Therefore, we can no
longer call one of our js functions in a link that opens a new tab.

For example, the following will now open a blank page and then fail
to load "a" because `voteSite` is not defined in that tab's context:

  <a href="javascript:voteSite(a,b)" target="_blank">

See https://bugs.chromium.org/p/chromium/issues/detail?id=992244

The fix to our voting links is to use a normal href link for the
external site (with `target="_blank"`), and then add an `onclick`
event to redirect the current page (if necessary, i.e. if free
turns are available):

  <a href="a" onclick="window.location='b'" target="_blank">

The real `onclick` event is only slightly more complicated than this.
We need to update the `window.location` _after_ we have opened the
external URL in a new tab. Therefore, the guts of `onclick` must be
wrapped in a `setTimeout`.

We change `VoteSite::getLinkUrl` to only return the external URL
and add `VoteSite::getSN`, which returns the internal SN or false.

Additional changes:
* Refactored the logic deciding whether or not free turns can be
  awarded into a private function `VoteSite::freeTurnsReady`.
* Moved the `alwaysreward=1` query param for TWG into the URL used
  when free turns are available (since it does nothing otherwise).