Declare strict typing
Use the following declaration after the php opening tag to ensure
that type hints are followed strictly:

declare(strict_types=1);

Note that this has to be specified _in every file_ due to a design
decision by the PHP developers (they wanted it to be an opt-in
feature, for better or worse).

This was deferred until PHP 7.4 due to an issue in PHP 7.3 with the
lack of nullable types in the mysqli interface.