SmrPlayer: decrease PvE experience loss on death
The experience loss on PvE death is still rather severe, especially
force deaths. We reduce the PvE death exp loss as follows to bring it
in line with PvP death exp loss, which is 15% on average (5-25%).

Forces:  25% -> 15%
Ports:   20-28% -> 15-20%
Planets: 20-27% -> 15-20%

The port/planet numbers are set to have an average slightly higher
than 15% since death to ports typically only happens against higher
level ports (so the real average is likely closer to 15% than 17.5%).

That said, I feel like the same can be said for PvP deaths (whose real
average likely sits below 15% since higher level players are more
likely to win PvP battles), but it is better to make a smaller change
now and tweak it further later if necessary.