SocialLogin: move authentication logic into class
Create a new static method `SocialLogin::getLoginUrl`, which contains
the logic that used to be in login_social_processing.php.

We switch universally to the class constant names of the social login
types, using the uppercase "Facebook" and "Twitter" since those are
already inserted into the database (i.e. only the query parameter is
modified).