login_processing.php: redirect for display pages
We don't want to display pages from processing containers, so when we
detect that a social login is not linked to an account, we will now
redirect to the login_social_create.php page.

Part of the problem was that if the user refreshed this display page,
it would redo all the processing, including the authentication with
the social platform, which would throw an exception (you are not
allowed to authenticate more than once with the same token), e.g.:

Facebook\Exceptions\FacebookSDKException: Cross-site request forgery
validation failed. Required param "state" missing from persistent data.

This may fix some of the bug reports we've gotten about social logins.