npc.php: don't use a hard-coded game ID
Put active NPCs to work for all enabled games instead of hard-coding
the `NPC_GAME_ID` in the NPC configuration file (which needed to be
manually updated each game). This is the simplest way to avoid manual
intervention for NPCs with each new game.

This new scheme uses the SmrSession to track which NPC is working,
rather than using globals in npc.php.

NOTE: This scheme may not work correctly if multiple games are active,
one of which is a non-Default game (because it may autoload the wrong
version of libraries).