ChessGame: get gameID from database not Session
In 8b55a57e0fd, we added a `game_id` field to the `chess_game` table.
Therefore, we no longer need to use the SmrSession to get the gameID
for the ChessGame class; we can simply use the value in the database.

The usage of `NPC_GAME_ID` in npc/chess.php was actually incorrect
in the case where there are multiple active games, because it would
associate all ongoing chess games with that game ID instead of the ID
in the database.