Lose relations when getting caught stealing
The low-probability event of getting caught stealing goods from a port
will now result in the same relations loss as a bad bargain.

It probably should be higher, given the severity of the offense, but
since it is unavoidable (whereas a bad bargain is avoidable), it should
not be overly punitive.