npc.php: fix handling of illegal goods in plotToFed
This function was only dumping illegal goods when it was in a sector
that offered federal protection. Since dumping cargo is not allowed
in Fed sectors, this was causing an error, leaving NPCs unprotected
with illegal goods on board.

To fix this problem, we dump illegal goods before we plot to Fed.
For simplicity and safety, we do this unconditionally (regardless of
the reason the NPC is going to Fed).

If the NPC tries to plot to Fed with illegal goods on board while in
a Fed sector already, then this error will still occur, and I don't
think that it's worth coding a workaround into the AI. We can always
simply delete the NPC cargo if we must (NPCs don't necessarily need
to follow the rules the way players do).