SmrPlayer: createBounty now returns the bounty
Callers of `createBounty` were expecting it to return the bounty
it created. Since it didn't have a return value, it returned `null`,
and PHP does not warn when trying to index a non-array variable
(this will be fixed in PHP 7.4).

Now `createBounty` will return the bounty it creates, and we add
type hinting to this function for completeness.