SmrPlayer: use intval for type correctness
Since `floor` returns a float, we can use `intval` instead to return
an int and avoid a warning about type-correctness when passing the
value to `SmrPlayer::increaseCurrentBountyAmount`. (The value is not
changed because we are only dealing with positive numbers.)

The same is true for `round`, except any difference in the resulting
value doesn't matter because we're just computing the bounty amount
for podding a port raider.