login_create.php: move recaptcha script to end of body
I thought that while testing the templatization of this page the
recaptcha wasn't showing up unless the recaptcha `<script>` was at
the top of the page.

This appears not to be true upon further testing, so move it to the
bottom of the page where it won't break page rendering (see #744).