Prevent overwriting planet mounted weapons
When adding a mounted weapon, it is possible that a weapon has been
added to the selected slot in the time between when the "Defense" page
was rendered and when the "Transfer" button was clicked.

To mitigate this (rare) scenario, the processing script will now raise
an error if the selected slot is already occupied.