Fix HoF subcategories overflowing table
When the HoF categories page was converted to a template, the
subcategories were joined with non-breaking spaces (presumably to try
to achieve a decent spacing between the buttons). However, this caused
the buttons to overflow the table.

To fix this, we switch from a non-breaking space to a regular space
(it needs to be the HTML entity `&#32;` otherwise it will be removed
by the whitespace-stripping function).