forces_drop_processing.php: don't error on empty
This fixes an error that players would get when they leave an input
field on the forces_drop.php page blank.

In 4452a2dba9f, this script was modified to simplify the checks on
the input fields with the addition of the "number" HTML5 input types.
However, empty strings are valid HTML5 "number" type values, and this
was not accounted for in the PHP validation logic.

Add in an explicit check to set the value to "0" if the input value
is the empty string.