Fix Galaxy Map showing traders on planets (#778)
Significant optimizations were made to the Galaxy Map in 8379956c5,
including a new function `SmrPlayer::getGalaxyPlayers`, which used a
single database query to cache the results for all subsequent calls
to `SmrPlayer::getSectorPlayers` (to avoid making a separate database
query for each sector in the galaxy).

The optimization is fine in principle, but `getGalaxyPlayers` was
caching the wrong results because the query was not identical to the
one in `getSectorPlayers`. Most importantly, it didn't exclude players
on planets, and so the Galaxy Map was incorrectly displaying them
(even enemy players, which had serious gameplay implications).

The fix is simply to make sure the queries between the two functions
in SmrPlayer are consistent.