Fix handling of SmrSession in npc.php
When switching NPCs, it is more consistent to destroy and re-init the
SmrSession. Otherwise, there is no good way to switch accounts within
the same session, which can result in confusing messages, e.g.:

2019-10-16 22:37:04 - Released NPC: npc2
2019-10-16 22:37:04 - Choosing new NPC
2019-10-16 22:37:04 - No free NPCs
2019-10-16 22:37:04 - Exiting NPC script.
2019-10-16 22:37:04 - Failed to release NPC: npc2

This example was a result of not updating the SmrSession account after
successfully releasing the NPC, and then trying to release the same
NPC again during normal shutdown.

We now also do not set cookies in `SmrSession::init` (in a reversal
of the decision in abc87d52682d2), because otherwise a re-init will
trigger a warning about output already started.