Fix unescaped HTML special characters in HoF names (#781)
Add a new function `SmrAccount::getHofDisplayName`, which returns the
HoF name in a form that is suitable for display (i.e. all HTML special
characters are escaped). The existing `SmrAccount::getHofName` returns
only the raw HoF name.

Most notably, there are some HoF names that use the `&`, which when
displayed unescaped causes DOMDocument to issue a warning.

Also fixed the link to the user's album page on the "Edit Photo" page.
It was missing the `nick` query string param added in 9e11f611990a5f5.
It is also now url-encoded.