config.php: always report all error levels
All operational modes will use E_ALL reporting so that, in particular,
the E_STRICT and E_NOTICE messages are not suppressed in production.
Previously this was restricted to debug mode only.

Note that `E_STRICT` errors (included by `E_ALL`) are evaluated at
compile time, so depend on the `error_reporting` setting in php.ini.
These were only turned on with the "development" presets.