npc.php: raise an exception at all error levels
Raise an exception when any error (warning, notice, etc.) is
encountered by using the custom error handler, i.e. the one we use for
beta. This has two benefits:

1. It makes warnings and notices more visible, since they would likely
   get lost in the flood of NPC log messages otherwise.

2. It attempts to convert every error into an exception, which is
   important for making sure that we shut down the NPC as cleanly as
   possible after an error. It is especially useful for "recoverable
   fatal errors", which just immediately exit from the default error
   handler (and leave the NPC in a broken state).

Note that not all errors can be converted into exceptions, but this
should catch the vast majority of cases that we might expect.