SmrWeapon: improve edge cases of checkHit
The WeightedRandom code changes the weight in the wrong direction for
accuracies outside the range [0%, 100%]. To fix this, we limit the
input accuracy to this range in WeightedRandom->flipWeightedCoin.

However, even with this fix, it can be unpleasantly unexpected to see
a shot miss when you have a (non-weighted) accuracy of >= 100%. To
address this, we simply return early in SmrWeapon->checkHit if the
accuracy is outside the range (0%, 100%), skipping the weighting
altogether.