Improve types from database queries
Many of the errors caused by turning on type-strictness in #795 were
the result of getting content from the database as a string and then
using it like an int, float, or boolean.

We can use the `get{Int,Float,Boolean}` methods of MySqlDatabase
instead of `getField` (which returns a string) to improve the type-
correctness of our database query results.

Not every change made here is necessary to avoid an exception, but it
does make sense to try to be consistent about our data types.