Remove dangling-ifs
It is safer to use explicit braces when a conditional extends onto
multiple lines. (Without them, you have to trust that the indentation
correctly implies the scope of the conditional).

This is done manually because Scrutinizer's auto-formatting doesn't
correctly fix dangling conditionals (it adds too much indentation).

git grep "if\s*\(.*\)\s*$"

VIM: /if\s*(.*)\s*$