Remove gadgets
The implementation of gadgets was incomplete and could not be used in
its current state. With the addition of more strict type-checking, we
were encountering errors in the no-op check for installed gadgets when
firing weapons. The simplest fix is to remove gadgets completely.

If we decide to include something like gadgets in the game, it is very
likely that the gadget code would be entirely rewritten anyway.

This change does make `getModifiedDamage` a no-op function, which we
could remove; however, for now just update the interface to address
a Scrutinizer warning about `$weaponPlayer` being unused.