Add rounding functions that return integer types
The functions `round`, `floor`, and `ceil` return a float type, even
though the returned value is a whole number. With strict types enabled,
we can not pass the result of one of these functions as an argument
that is type-hinted as an integer.

We add wrappers for these functions: `IRound`, `IFloor`, and `ICeil`,
which cast the return value as an int type.

NOTE: While `intval` is another method of converting to int types, we
want to avoid it, where possible, since it casts _any_ type to an int.
If we know the input is a float, then we should use `IFloor` instead.