change power checking behavior
a devastator, with 26 power, should be able to mount 8 level 3 weapons for a total of 24 power.  current code checking does not permit that.  it treats the first slot as up to level 5, next two slots as up to 4, next 3 up to 3, and so on.

either the concept of ship power is misleading or this is unintended behavior.  probably easier to sort it out this way.