RouteGenerator: fix sorting routes by multipliers
The multipliers were being used as array keys; however, when a float
is used as an array key, it is implicitly converted to an integer!
As a result, we were not correctly sorting routes by multipliers.

To fix this, we convert the multipliers to strings so that they can
be used as array keys, and then force a SORT_NUMERIC sorting scheme.