OneWayRoute: fix money multiplier calculation
The transactions in the calculation were backwards from what the
ports actually offered. Make it clear that the transactions are
from the perspective of the player, not the port.

Also adjusted the supply / relations factors in the calculation
for consistency with SmrPort::getIdealPrice. (Ideally, we would
move the formula to a static method so that we do not need to
manually keep the SmrPort and OneWayRoute classes synchronized.)