RouteGenerator: fix bug on repeated calls
The `$dontAddWorseThan` static variable was not being reset, so when
the route generation functions were called multiple times, it was
incorrectly truncating the results.

Add an initialization function to consistently set up the static
variables (though this class probably should not be using static
public methods and storage variables at all).