SmrPort: fix "Levels Raided" HoF credit
When razing was added back in 2013 (38e4b44aab), the refactoring of
the attacker list into `getAttackersToCredit` left a dangling database
reference in `creditCurrentAttackersForKill`, causing the following
notice (new in PHP 7.4) to be emitted:

Notice: Trying to access array offset on value of type null

As a result, the "Levels Raided" Hall of Fame category was not being
populated.

The simplest fix is to use the current level of the port, rather than
the level the port was at the last time each player attacked it (as it
is stored in the `player_attacks_port` table). Regardless of which way
the stats are awarded, the impact should be completely negligible.