Menu: fix database error on Alliance Roster
If the player viewing the Alliance Roster did not have an alliance,
then the query to check the `send_alliance_msg` permission would fail
because the record would not exist.

We still perform the query, but now we do so with the correct alliance
role, and we only access the `send_alliance_msg` field if the record
exists (i.e. if the player is in or has a treaty with this alliance).