SmrAlliance: fix handling of description
The `description` field in the `alliance` table is nullable, so we do
not want to pass its value directly into a function that expects a
string (i.e. `stripslashes` or `real_escape_string`).

This property was escaped before it was inserted into the database via
a call to `htmlentities` in `setAllianceDescription`, so the post-query
escaping was not even necessary. However, the escaping should be the
responsibility of display methods, so we move the escaping logic into
`getDescription`. (This also prevents the escaped content from being
displayed when editing the description. We remove `nl2br` because it
was adding extra line breaks.)

NOTE: The same is true for the `mod` field, so we change it as well,
for consistency, even though it is not a nullable field.