ChessGame: fix array access on bool
When a non-castling move is made, the castling variable is defined as
`false` instead of an array. In PHP 7.4, a warning is emitted when you
try to index a variable that is not an array:

> PHP Notice:  Trying to access array offset on value of type bool

This caused ajax updates to break for any non-castling move.

To fix this, we first check if the variable is `false` or an array,
and then proceed accordingly.