Fix player vs. player drones strict type error
The formula for calculating drone accuracy in player vs. player combat
has a factor of the form:

rand(player_level / 2, player_level)

Since `rand` only takes integer arguments, this raises an error with
strict typing enabled. If we cast the first argument to an int using
`IFloor`, the resulting range is unchanged.