SmrSession: truly destroy sessions
In the distant past, we may have used rows in the `active_session`
table to display how many players were online, and this is likely why
`SmrSession::destroy` would retain non-expired sessions, updating the
other columns to empty values and then redefining the important static
SmrSession properties (session, game, and account IDs).

This retention, while unnecessary, was functional for regular player
access to the game. However, for NPCs, setting `$session_id` to the
empty string was a problem because it would cause subsequent calls to
`SmrSession::init` (when we switch to a new NPC) to not generate a new
session, using the empty string session ID, which caused duplicate
primary key errors.

Now, we truly destroy the session, deleting the `active_session` row
and setting `$session_id = null` instead of the empty string. For the
issue with NPCs, only the latter change is needed.