SmrGame: use === when checking equality
There was an issue with strict typing where setting an SmrGame
attribute to the empty string would not override the default value
of `null` because we were comparing `'' == null`, which is true.
As a result, we were passing `null` to `escapeString`, which threw
an exception.

To fix this, we want to use the typed comparison operator `===`,
because `'' === null` is false.