Make submitStyle links look like submit input elements
This change fixes the box spacing around links with the `submitStyle`
class so that they are styled the same way as submit input elements.

In Chrome, the user agent stylesheet specifies `padding` for submit
input elements as 1px (top/bottom) and 6px (left/right). Perhaps this
was different (0px and 8px) in the past or for other browsers. We
should probably set this explicitly, to ensure it is consistent with
the `submitStyle` class, but for now it is sufficient to assume a
recent Chrome version.

The `border-spacing` is a more subtle issue. We set it to 0 for a lot
of tables in our CSS, and it seems to be inherited by `submitStyle`
links contained in those tables. However, if the link is inside a
default table, then its `border-spacing` gets reset to the user agent
stylesheet default of 2px. We set it explicitly for `submitStyle` to
avoid this issue.