Break account_edit.php into two pages
The logic can be significantly simplified by having a separate page
for searching for a player and for displaying/editing the selected
player.

This allows us to use the Request methods (without default values)
instead of SmrSession::getRequestVar (with default values), which
is safer and more explicit about the expectations for processing.