shop_goods_processing.php: fix stealing exploit
By falsifying form data, players could steal goods that a port buys
(instead of goods that it sells).

Fix this exploit by throwing an exception if the player tries to
steal a good that the port doesn't sell.

NOTE: We use an exception instead of `create_error` because we want
to be alerted when players are falsifying form data.