SmrPlayer: fix type issue with bounty functions
Raise an exception when calling `getBounty` with a `$bountyID` that
does not exist instead of returning false. This allows us to guarantee
that the return value is an array. Use `hasBounty` first if there is
any danger that the `$bountyID` does not exist.

Fixes the Scrutinizer error:

It seems like $bounty can also be of type false; however, parameter
$bounty of AbstractSmrPlayer::setBounty() does only seem to accept
array, maybe add an additional type check?