Refactor Rankings pages
The player and alliance rankings pages had two versions of almost
identical database queries. This is susceptible to unintentional
desynchronization.

To fix this, we write out the query only once in a function, and
allow parameters to specify the limits of the query.

Also default to centered on the current sector in Sector Kills,
rather than defaulting to the top 10 (to match what we do on all
the other ranking pages).