contact_processing.php: don't spoof player e-mails
When sending e-mails from the Contact Form to our support mailbox, we
were using the player's e-mail address as the sender. Since we don't
have any authority to send e-mail on behalf of an address that we do
not own (this is called "spoofing"), send as "contact@smrealms.de"
instead.

Put the player's e-mail as the "Reply To" address to make it easier
for admins to respond to the player, if necessary.