account_edit.php: make player names placeholders
By using placeholders for the existing player names, if the names are
unchanged, then the submitted names are empty. And when the submitted
names are empty, we avoid a significant amount of processing that is
usually an expensive no-op (grabbing the SmrPlayer for this account
from every game and making sure the name hasn't changed...).

Other benefits:

* Placeholders make it obvious which player names were modified, since
  they are displayed as grey text instead of white.

* If you change your mind and delete an input name, the placeholder
  will automatically reappear.