Add Request class (#824)
Data in the `$_REQUEST` superglobal is stored as strings and is not
sanitized in any way. By demanding a specific type when we query the
data, we a) facilitate the use of strict typing throughout the code
and b) can skip a lot of sanity checks that are automatically taken
care of by the type conversion.

This synergizes with HTML5 form validation, which makes the implicit
type conversion in the Request class less destructive (because we
should only get unexpected data types in the cases where players are
circumventing the HTML5 validation to falsify form data).

The Request class is designed for use in processing pages. Most of
the time, display pages should be using `SmrSession::getRequestVar`.

We switch from POST -> REQUEST in many places because they are both
user submitted, so there is no reason to demand one or the other.