Add Menu::headquarters
This replaces the `create_{hq,ug}_menu` functions with a new method
in the `Menu` class.

As a result, Menu.class.php is the only place where the `create_menu`
function is now used, so we can move it out of smr.inc and into the
Menu class (to keep all of the menu logic consolidated).

Since the "Semi Wars" game type has different headquarters menus, we
need to implement this using an abstract base class `AbstractMenu`.