Use getColouredRaceName second argument consistently
When the second argument is `true`, the race name becomes a link to
the given race's council page.

In some cases, this just simplifies the code (by generating the link
automatically instead of creating it manually). In others, it adds a
link to the race name where there was none before, e.g.:

* Alliance Roster (for player race)
* Weapon Shop (for weapon race)