Update to traefik v2
There is a major design change between v1.7 and v2+: "backends and
frontends are replaced with routers and middlewares". As a result,
all of our traefik labels need to be updated; in particular, the host
rule and http path manipulation are no longer combined and need to be
specified separately with the `rule` and `middleware` labels.

Many of our services (traefik, api-docs, portainer, and pma) require
a trailing slash at the root entrypoint to work correctly. To do this
in traefik v2, we define the `add-slash` middleware. We also need to
chain this with the `strip-pathprefix` middleware (which assumes the
role that `PathPrefixStrip` had in v1.7).

Some http paths needed to change:

* `/traefik` is now accessed at `/dashboard` due to a limitation with
  customizing the dashboard absolute URI in v2.

* `/api` (from the api-docs service) is now accessed at `/docs` due to
  the traefik v2 dashboard requiring both the `/api` and `/dashboard`
  endpoints.

I've also taken this opportunity to clean up some unnecessary traefik
labels in the docker-compose file:

* All http services being served by traefik will be connected to via
  the `frontend` network, so make it the default.

* Port does not need to be specified if the docker container exposes a
  single port.

* Don't specify `passHostHeader=true` because it is the default in the
  http service configuration.

* Set `exposedByDefault=true` for the docker provider so that only
  services that explicitly set `traefik.enable=true` are included in
  the routing configuration.